requirements.txt
africastalking==1.2.3
asgiref==3.3.1
astroid==2.5.1
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
contextlib2==0.6.0.post1
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.1
pylint==2.7.2
pytz==2021.1
requests==2.25.1
schema==0.7.4
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
